+++ benchmark/benchmark_test.go
@@ -14,27 +14,27 @@ import (
 // BenchmarkPutsWithSecondaryV1  ... Takes in an async worker count and profiles blob insertions using
 // constant blob sizes in parallel. Exercises V1 code pathways
 func BenchmarkPutsWithSecondaryV1(b *testing.B) {
-	testConfig := testutils.NewTestConfig(testutils.MemstoreBackend, false)
-	putsWithSecondary(b, testConfig)
+	testCfg := testutils.NewTestConfig(testutils.MemstoreBackend, false)
+	putsWithSecondary(b, testCfg)
 }
 
 // BenchmarkPutsWithSecondaryV2  ... Takes in an async worker count and profiles blob insertions using
 // constant blob sizes in parallel. Exercises V2 code pathways
 func BenchmarkPutsWithSecondaryV2(b *testing.B) {
-	testConfig := testutils.NewTestConfig(testutils.MemstoreBackend, true)
-	putsWithSecondary(b, testConfig)
+	testCfg := testutils.NewTestConfig(testutils.MemstoreBackend, true)
+	putsWithSecondary(b, testCfg)
 }
 
-func putsWithSecondary(b *testing.B, testConfig testutils.TestConfig) {
-	testConfig.UseS3Caching = true
+func putsWithSecondary(b *testing.B, testCfg testutils.TestConfig) {
+	testCfg.UseS3Caching = true
 	writeThreadCount := os.Getenv("WRITE_THREAD_COUNT")
 	threadInt, err := strconv.Atoi(writeThreadCount)
 	if err != nil {
 		panic(fmt.Errorf("Could not parse WRITE_THREAD_COUNT field %w", err))
 	}
-	testConfig.WriteThreadCount = threadInt
+	testCfg.WriteThreadCount = threadInt
 
-	tsConfig := testutils.BuildTestSuiteConfig(testConfig)
+	tsConfig := testutils.BuildTestSuiteConfig(testCfg)
 	ts, kill := testutils.CreateTestSuite(tsConfig)
 	defer kill()
